refactor Engine
